+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-04-22 19:27:44 +0300 using RuboCop version 0.67.2.
+# on 2019-04-22 19:33:17 +0300 using RuboCop version 0.67.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -15,7 +15,7 @@ Lint/UriEscapeUnescape:
 Metrics/AbcSize:
   Max: 23
 
-# Offense count: 12
+# Offense count: 13
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
@@ -49,6 +49,7 @@ Naming/AccessorMethodName:
   Exclude:
     - 'lib/teamlab/modules/calendar.rb'
     - 'lib/teamlab/modules/community.rb'
+    - 'lib/teamlab/modules/community/community_bookmarks.rb'
     - 'lib/teamlab/modules/crm.rb'
     - 'lib/teamlab/modules/files.rb'
     - 'lib/teamlab/modules/group.rb'
@@ -93,7 +94,7 @@ Style/Documentation:
     - 'lib/teamlab/modules/portals.rb'
     - 'lib/teamlab/modules/project.rb'
     - 'lib/teamlab/modules/settings.rb'
-    - 'lib/teamlab/request.rb'g
+    - 'lib/teamlab/request.rb'
     - 'lib/teamlab/response.rb'
 
 # Offense count: 4
