+++ lib/teamlab/modules/community.rb
@@ -1,5 +1,8 @@
+require_relative 'community/community_bookmarks'
 module Teamlab
   class Community
+    include CommunityBookmarks
+
     def initialize
       @request = Teamlab::Request.new('community')
     end
@@ -50,54 +53,6 @@ module Teamlab
 
     def delete_post(post_id)
       @request.delete(['blog', post_id.to_s])
-    end
-
-    def get_all_bookmarks
-      @request.get(%w[bookmark])
-    end
-
-    def get_all_bookmark_tags
-      @request.get(%w[bookmark tag])
-    end
-
-    def get_bookmark(id)
-      @request.get(['bookmark', id.to_s])
-    end
-
-    def get_bookmarks_added_by_me
-      @request.get(%w[bookmark @self])
-    end
-
-    def get_my_favourite_bookmarks
-      @request.get(%w[bookmark @favs])
-    end
-
-    %w[day week month year].each do |e|
-      define_method("get_top_of_#{e}_bookmarks") { @request.get(['bookmark', 'top', e.to_s]) }
-    end
-
-    def get_bookmarks_by_tag(tag)
-      @request.get(['bookmark', 'tag', tag.to_s])
-    end
-
-    def get_recently_added_bookmarks
-      @request.get(%w[bookmark top recent])
-    end
-
-    def get_bookmark_comments(bookmark_id)
-      @request.get(['bookmark', bookmark_id.to_s, 'comment'])
-    end
-
-    def search_bookmarks(queue)
-      @request.get(['bookmark', '@search', queue.to_s])
-    end
-
-    def add_bookmark(url, title, options = {})
-      @request.post(%w[bookmark], { url: url, title: title }.merge(options))
-    end
-
-    def add_comment_to_bookmark(bookmark_id, content, options = {})
-      @request.post(['bookmark', bookmark_id.to_s, 'comment'], { content: content }.merge(options))
     end
 
     def get_all_events
